+++ vm/src/tvm/language_model.cpp
@@ -10,7 +10,7 @@
 
 #include "../chat_manager.hpp"
 #include "../model_cache.hpp"
-#include "tokenizer.hpp"
+#include "../tokenizer.hpp"
 #include "tvm_model.hpp"
 
 #include <iostream>
